Be explicit about scalar_checks in a few places.
The goal here is to either be explicit about or to remove implicit scalar_checks completely from Declarations.cwrap.

I'm not sure what the correct strategy is yet, this is just an initial obvious set.

There is no behaviorial change, but this will invoke maybe_zero_dim(false), which will immediately return.

ghstack-source-id: d5fa46ed42c73c865b49a4c2e6e16659e462827e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/29759